refactor(integration/spring): rewrite SpringKotlinIrEmitter as an IrTransformer
SpringKotlinIrTransformer applies the same Spring MVC mapping
annotations and WirespecNativeHints generation as the deleted
SpringKotlinIrEmitter, but as an IrTransformer that can be registered
on any Kotlin IrEmitter via the new plugin transformers parameter.
Endpoint files are matched by their namespace extending
Wirespec.Endpoint, and the hints file is now appended as an IR File
